+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/Checkbox/CheckboxIndeterminateConfiguration.swift
@@ -34,12 +34,24 @@ final class CheckboxIndeterminateConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
+    @Published var isReadOnly: Bool {
+        didSet { updateCode() }
+    }
+
+    var accessibilityLabel: String {
+        (isError ? "app_components_common_error_a11y" : "app_components_checkbox_hint_a11y")
+            .localized()
+    }
+
     // MARK: Initializer
 
     override init() {
+        enabled = true
         indicatorState = .selected
         isError = false
+        isReadOnly = false
         enabled = true
+        super.init()
     }
 
     deinit {}
@@ -49,7 +61,7 @@ final class CheckboxIndeterminateConfigurationModel: ComponentConfiguration {
     override func updateCode() {
         code =
             """
-            OUDSCheckboxInterminate(selection: $selection\(isErrorPattern))
+            OUDSCheckboxInterminate(selection: $selection\(isErrorPattern)\(isReadOnlyPattern))
             \(disableCodePattern)
             """
     }
@@ -65,6 +77,10 @@ final class CheckboxIndeterminateConfigurationModel: ComponentConfiguration {
             ""
         }
     }
+
+    private var isReadOnlyPattern: String {
+        isReadOnly ? ", isReadOnly: true" : ""
+    }
 }
 
 // MARK: - Checkbox Indeterminate Configuration View
@@ -78,14 +94,18 @@ struct CheckboxIndeterminateConfiguration: View {
     var body: some View {
         VStack(alignment: .leading, spacing: theme.spaces.fixedNone) {
             OUDSSwitchItem("app_common_enabled_label", isOn: $configurationModel.enabled)
-                .disabled(configurationModel.isError)
+                .disabled(configurationModel.isReadOnly || configurationModel.isError)
+
+            OUDSSwitchItem("app_components_common_readOnly_label", isOn: $configurationModel.isReadOnly)
+                .disabled(!configurationModel.enabled || configurationModel.isError)
 
             OUDSSwitchItem("app_components_common_error_label", isOn: $configurationModel.isError)
-                .disabled(!configurationModel.enabled)
+                .disabled(!configurationModel.enabled || configurationModel.isReadOnly)
 
             OUDSChipPicker(title: "app_components_common_selection_label",
                            selection: $configurationModel.indicatorState,
                            chips: OUDSCheckboxIndicatorState.chips)
+                .disabled(!configurationModel.enabled || configurationModel.isReadOnly)
         }
     }
 }