+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/Checkbox/CheckboxPage.swift
@@ -46,8 +46,9 @@ private struct CheckboxDemo: View {
         HStack(alignment: .center) {
             Spacer()
             OUDSCheckbox(isOn: $configurationModel.indicatorState,
-                         accessibilityLabel: "app_components_checkbox_hint_a11y".localized(), // No LocalizedStringKey inference in the component
-                         isError: configurationModel.isError)
+                         accessibilityLabel: configurationModel.accessibilityLabel,
+                         isError: configurationModel.isError,
+                         isReadOnly: configurationModel.isReadOnly)
                 .disabled(!configurationModel.enabled)
                 .accessibilityIdentifier(A11YIdentifiers.componentCheckbox)
             Spacer()