+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/CheckboxItem/Indeterminate/CheckboxItemIndeterminatePage.swift
@@ -21,7 +21,7 @@ struct CheckboxItemIndeterminatePage: View {
     @StateObject private var configurationModel: IndicatorControlItemConfigurationModel
 
     init() {
-        let model = IndicatorControlItemConfigurationModel(componentInitCode: "OUDSCheckboxItemIndeterminate(selection: $selection,")
+        let model = IndicatorControlItemConfigurationModel(componentInitCode: "OUDSCheckboxItemIndeterminate(selection: $selection")
         _configurationModel = StateObject(wrappedValue: model)
     }
 
@@ -43,14 +43,15 @@ private struct CheckboxItemIndeterminateDemo: View {
     var body: some View {
         OUDSCheckboxItemIndeterminate(selection: $configurationModel.selection,
                                       label: configurationModel.labelText,
-                                      helper: configurationModel.helperText,
+                                      description: configurationModel.descriptionText,
                                       icon: icon,
                                       flipIcon: configurationModel.flipIcon,
                                       isReversed: configurationModel.isReversed,
                                       isError: configurationModel.isError,
                                       errorText: configurationModel.errorText,
                                       isReadOnly: configurationModel.isReadOnly,
-                                      hasDivider: configurationModel.hasDivider)
+                                      hasDivider: configurationModel.hasDivider,
+                                      constrainedMaxWidth: configurationModel.constrainedMaxWidth)
             .disabled(!configurationModel.enabled)
             .padding(.all, theme.spaces.fixedMedium)
             .designToolboxColoredSurface(false)