+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/TextInput/TextInputConfigurationView.swift
@@ -24,7 +24,7 @@ final class TextInputConfigurationModel: ComponentConfiguration {
     private let defaultLabel = String(localized: "app_components_common_label_label")
     private let defaultHelperText = String(localized: "app_components_common_helperText_label")
     private let defaultPlaceholderText = String(localized: "app_components_textInput_placeholder_label")
-    private let defaultErrorText = String(localized: "app_components_common_errorText_label")
+    private let defaultErrorText = String(localized: "app_components_common_errorMessage_label")
 
     // MARK: Published properties
 
@@ -83,6 +83,10 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
+    @Published var constrainedMaxWidth: Bool {
+        didSet { updateCode() }
+    }
+
     @Published var status: OUDSTextInput.Status {
         didSet { updateCode() }
     }
@@ -103,7 +107,9 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         text = ""
         helperLinkText = ""
         isOutlined = false
+        constrainedMaxWidth = false
         status = .enabled
+        super.init()
     }
 
     deinit {}
@@ -114,7 +120,7 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         // swiftlint:disable line_length
         code =
             """
-            OUDSTextInput(\(labelPattern)\(textPattern)\(placeholderPattern)\(prefixPattern)\(suffixPattern)\(leadingIconPattern)\(flipLeadingIconPattern)\(trailingActionPattern)\(helperTextPattern)\(helperLinkPattern)\(outlinedPattern)\(statusPattern))
+            OUDSTextInput(\(labelPattern)\(textPattern)\(placeholderPattern)\(prefixPattern)\(suffixPattern)\(leadingIconPattern)\(flipLeadingIconPattern)\(trailingActionPattern)\(helperTextPattern)\(helperLinkPattern)\(outlinedPattern)\(constrainedMaxWidthPattern)\(statusPattern))
             """
         // swiftlint:enable line_length
     }
@@ -165,6 +171,10 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         isOutlined ? ", isOutlined: true" : ""
     }
 
+    private var constrainedMaxWidthPattern: String {
+        constrainedMaxWidth ? ", constrainedMaxWidth: true" : ""
+    }
+
     private var statusPattern: String {
         status != .enabled ? ", status: \(status.technicalDescription)" : ""
     }
@@ -182,6 +192,8 @@ struct TextInputConfigurationView: View {
         VStack(alignment: .leading, spacing: theme.spaces.fixedMedium) {
             VStack(alignment: .leading, spacing: theme.spaces.fixedNone) {
                 OUDSSwitchItem("app_components_common_outlined_label", isOn: $configurationModel.isOutlined)
+
+                OUDSSwitchItem("app_components_common_constrainedMaxWidth_label", isOn: $configurationModel.constrainedMaxWidth)
 
                 OUDSSwitchItem("app_components_textInput_leadingIcon_label", isOn: $configurationModel.leadingIcon)