+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/TextInput/TextInputElement.swift
@@ -35,7 +35,7 @@ private struct TextInputIllustration: View {
     @State private var text: String = ""
 
     var body: some View {
-        OUDSTextInput(label: String(localized: "app_components_common_label_label"),
+        OUDSTextInput(label: "Label",
                       text: $text,
                       helperText: String(localized: "app_components_textInputHelperText_label"))
     }