+++ DesignToolbox/DesignToolboxSnapshotsTests/__Snapshots__/Tokens/Test Cases/TokensSpaceSnapshotsTestsTestCase.swift
@@ -75,6 +75,34 @@ open class TokensSpaceSnapshotsTestsTestCase: XCTestCase {
         }
     }
 
+    /// Tests all padding inset `Inset` spaces by capturing their snapshots.
+    /// - Parameters:
+    ///   - theme: Theme used for rendering tokens (e.g. `OrangeTheme`).
+    ///   - interfaceStyle: The user interface style (light or dark) for which to test the colors.
+    @MainActor func testInset(theme: OUDSTheme, interfaceStyle: UIUserInterfaceStyle) {
+
+        // Iterate through all background color cases defined in NamedSpace.Inset
+        for namedToken in NamedSpace.Inset.allCases {
+            // Generate the illustration for the specified space token using the spacePage instance
+            let illustration = OUDSThemeableView(theme: theme) {
+                SpaceTokenVariant(namedSpaceToken: namedToken) { token in
+                    InsetSpaceProperty.Illustration(token: token)
+                }
+                .background(theme.colors.bgPrimary.color(for: interfaceStyle == .light ? .light : .dark))
+            }
+
+            // Create a unique snapshot name based on the current mode (light or dark) and the color's raw value
+            let testName = "test_\(theme.name)Theme_\(interfaceStyle == .light ? "Light" : "Dark")"
+            let name = namedToken.rawValue
+
+            // Capture the snapshot of the illustration with the correct user interface style and save it with the snapshot name
+            assertIllustration(illustration,
+                               on: interfaceStyle,
+                               named: name,
+                               testName: testName)
+        }
+    }
+
     /// Tests all padding inline `PaddingInline` spaces by capturing their snapshots.
     /// - Parameters:
     ///   - theme: Theme used for rendering tokens (e.g. `OrangeTheme`).
@@ -102,35 +130,7 @@ open class TokensSpaceSnapshotsTestsTestCase: XCTestCase {
         }
     }
 
-    /// Tests all padding inset `PaddingInset` spaces by capturing their snapshots.
-    /// - Parameters:
-    ///   - theme: Theme used for rendering tokens (e.g. `OrangeTheme`).
-    ///   - interfaceStyle: The user interface style (light or dark) for which to test the colors.
-    @MainActor func testPaddingInset(theme: OUDSTheme, interfaceStyle: UIUserInterfaceStyle) {
-
-        // Iterate through all background color cases defined in NamedSpace.PaddingInset
-        for namedToken in NamedSpace.PaddingInset.allCases {
-            // Generate the illustration for the specified space token using the spacePage instance
-            let illustration = OUDSThemeableView(theme: theme) {
-                SpaceTokenVariant(namedSpaceToken: namedToken) { token in
-                    PaddingInsetProperty.Illustration(token: token)
-                }
-                .background(theme.colors.bgPrimary.color(for: interfaceStyle == .light ? .light : .dark))
-            }
-
-            // Create a unique snapshot name based on the current mode (light or dark) and the color's raw value
-            let testName = "test_\(theme.name)Theme_\(interfaceStyle == .light ? "Light" : "Dark")"
-            let name = namedToken.rawValue
-
-            // Capture the snapshot of the illustration with the correct user interface style and save it with the snapshot name
-            assertIllustration(illustration,
-                               on: interfaceStyle,
-                               named: name,
-                               testName: testName)
-        }
-    }
-
-    /// Tests all padding stack `PaddingBlock` spaces by capturing their snapshots.
+    /// Tests all padding block `PaddingBlock` spaces by capturing their snapshots.
     /// - Parameters:
     ///   - theme: Theme used for rendering tokens (e.g. `OrangeTheme`).
     ///   - interfaceStyle: The user interface style (light or dark) for which to test the colors.